+++ src/main/php/net/stubbles/db/config/PropertyBasedDatabaseConfigReader.php
@@ -83,6 +83,17 @@ class PropertyBasedDatabaseConfigReader implements DatabaseConfigReader
     }
 
     /**
+     * returns list of available config ids
+     *
+     * @return  string[]
+     * @since   2.1.0
+     */
+    public function configIds()
+    {
+        return $this->readProperties()->getSections();
+    }
+
+    /**
      * checks whether database configuration for given id exists
      *
      * @param   string  $id