+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ------------------
 
   * added `net\stubbles\db\Database` to allow easier access to query results
+  * added `net\stubbles\db\ioc\ConnectionProvider::availableConnections()` to retrieve a list of available database connection ids
 
 
 2.0.0 (2013-11-10)